backend/requirements.txt
Flask==3.1.0
gunicorn==20.1.0
SQLAlchemy==2.0.37
python-dotenv==1.0.1
Werkzeug==3.1.3
requests==2.32.3
psycopg2-binary==2.9.10
Flask-SQLAlchemy==3.1.1